+++ test/unit/canvas_static.js
@@ -1624,6 +1624,25 @@
     canvas.viewportTransform = fabric.StaticCanvas.prototype.viewportTransform;
   });
 
+  QUnit.test('setViewportTransform calls objects setCoords', function(assert) {
+    var vpt = [2, 0, 0, 2, 50, 50];
+    assert.deepEqual(canvas.viewportTransform, [1, 0, 0, 1, 0, 0], 'initial viewport is identity matrix');
+    var rect = new fabric.Rect({ width: 10, heigth: 10 });
+    var rectBg = new fabric.Rect({ width: 10, heigth: 10 });
+    var rectOverlay = new fabric.Rect({ width: 10, heigth: 10 });
+    canvas.add(rect);
+    canvas.cancelRequestedRender();
+    canvas.backgroundImage = rectBg;
+    canvas.overlayImage = rectOverlay;
+    assert.deepEqual(rect.lineCoords.tl, new fabric.Point(0,0), 'rect linecoords are set for normal viewport');
+    assert.equal(rectBg.lineCoords, undefined, 'rectBg linecoords are not set');
+    assert.equal(rectOverlay.lineCoords, undefined, 'rectOverlay linecoords are not set');
+    canvas.setViewportTransform(vpt);
+    assert.deepEqual(rect.lineCoords.tl, new fabric.Point(50,50), 'rect linecoords are set');
+    assert.deepEqual(rectBg.lineCoords.tl,  new fabric.Point(0,0), 'rectBg linecoords are set');
+    assert.deepEqual(rectOverlay.lineCoords.tl,  new fabric.Point(0,0), 'rectOverlay linecoords are set');
+  });
+
   QUnit.test('getZoom', function(assert) {
     assert.ok(typeof canvas.getZoom === 'function');
     var vpt = [2, 0, 0, 2, 50, 50];